adding condition in frontend, implementing logic in backend + validation user auth
